# swagger.yaml
swagger: "2.0"
info:
  title: "dis-migration-service"
  description: "A Go API for data migration."
  version: 1.0.0
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
basePath: "/v1"
schemes:
  - https
...
paths:
  /migration-jobs:
    get:
      security:
        - Authorization: [migration:read]
      tags:
        - private
      summary: "Get list of migration jobs"
      description: "Lists all migration jobs in a paginated format."
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/limit"
        - $ref: "#/parameters/offset"
        - $ref: "#/parameters/state"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/MigrationJobList"
        400:
          description: "Invalid request parameter(s)"
          schema:
            $ref: "#/definitions/ErrorList"
        401:
          $ref: "#/responses/Unauthenticated"
        403:
          $ref: "#/responses/Forbidden"
        500:
          $ref: "#/responses/Error"

    post:
      security:
        - Authorization: [migration:create]
      tags:
        - private
      summary: "Create a migration job"
      description: "Creates a migration job using body data."
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - $ref: "#/parameters/dry_run"
        - $ref: "#/parameters/MigrationJobPostBody"
      responses:
        200:
          description: "Successful dry run response"
          schema:
            $ref: "#/definitions/MigrationJobPreview"
        202:
          description: "Successful task acceptance"
          schema:
            $ref: "#/definitions/MigrationJob"
        400:
          description: "Invalid request body or source title not found"
          schema:
            $ref: "#/definitions/ErrorList"
        401:
          $ref: "#/responses/Unauthenticated"
        403:
          $ref: "#/responses/Forbidden"
        409:
          description: "Already running job with these parameters"
          schema:
            $ref: "#/definitions/ErrorList"
        500:
          $ref: "#/responses/Error"
...
parameters:
  dry_run:
    in: query
    name: dry_run
    description: >
      A modifier for creation to see what would be created if done.
    type: boolean
    required: false
...
  limit:
    name: limit
    in: query
    description: "The number of items requested, defaulted to 10 and limited to 1000."
    type: integer
    required: false
    default: 10
  offset:
    in: query
    name: offset
    description: >
      The offset into the complete ordered set of items which satisfy the query, which specifies the 
      first item to return (starting at 0). Use this parameter as a pagination mechanism along with 
      the limit parameter."
    type: integer
    required: false
    default: 0
  state:
    in: query
    name: state
    description: >
      A filter for item state, specifies a subset of items to be returned"
    type: string
    enum: &STATE
      - submitted
      - in_review
      - approved
      - published
      - completed
      # Running states
      - migrating
      - publishing
      - post_publishing
      - reverting
      # Failed states
      - failed_post_publish
      - failed_publish
      - failed_migration
      - cancelled
    required: false
  MigrationJobPostBody:
    in: body
    name: body
    schema:
      type: object
      properties:
        source_id:
          type: string
        target_id:
          type: string
        type:
          $ref: "#/definitions/MigrationJobType"
      required:
        - source_id
        - type
...
responses:
  Unauthenticated:
    description: "Request is not authenticated"
    schema:
      $ref: "#/definitions/ErrorList"
  Forbidden:
    description: "Request is not authorised"
    schema:
      $ref: "#/definitions/ErrorList"
  Error:
    description: "Generic error response"
    schema:
      $ref: "#/definitions/ErrorList"
definitions:
...
  MigrationJob:
    type: object
    properties:
      job_number:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440000"
      last_updated:
        type: string
        format: date-time
        example: "2020-06-11T12:49:20+01:00"
      label:
        type: string
        description: "The title of the source dataset, automatically populated during job creation"
        example: "Consumer Price Index: All Items"
      links:
        type: object
        properties:
          events:
            description: "A link to all the events for the migration job"
            type: object
            properties:
              href:
                description: "A fully qualified URL to the migration job's events API response"
                type: string
                example: "https://api.beta.ons.gov.uk/v1/migration-jobs/550e8400-e29b-41d4-a716-446655440000/events"
          self:
            description: "A link to the individual migration job"
            type: object
            properties:
              href:
                description: "A fully qualified URL to the migration job API response"
                type: string
                example: "https://api.beta.ons.gov.uk/v1/migration-jobs/1"
              job_number:

                type: string
          tasks:
            description: "A link to all the tasks for the migration job"
            type: object
            properties:
              href:
                description: "A fully qualified URL to the migration job's tasks API response"
                type: string
                example: "https://api.beta.ons.gov.uk/v1/migration-jobs/550e8400-e29b-41d4-a716-446655440000/tasks"
      config:
        description: "Config for the job"
        $ref: "#/definitions/MigrationJobConfig"
      state:
        $ref: "#/definitions/MigrationState"
      type:
        $ref: "#/definitions/MigrationJobType"
  MigrationJobConfig:
    type: object
    properties:
      source_id:
        type: string
        example: "/economy/inflationandpriceindices"
      target_id:
        type: string
        example: "migration-series-id"
      type:
        $ref: "#/definitions/MigrationJobType"
  MigrationJobPreview:
    type: object
    properties:
      allOf:
        $ref: "#/definitions/MigrationJob"
      tasks:
        type: array
        items:
          $ref: "#/definitions/MigrationTask"
  MigrationTask:
    type: object
    properties:
      id:
        type: string
        example: 1
      last_updated:
        type: string
        format: date-time
        example: "2020-06-11T12:49:20+01:00"
      source:
        $ref: "#/definitions/MigrationTaskMetadata"
      state:
        $ref: "#/definitions/MigrationState"
      target:
        $ref: "#/definitions/MigrationTaskMetadata"
      type:
        $ref: "#/definitions/MigrationTaskType"
  MigrationTaskMetadata:
    type: object
    properties:
      id:
        type: string
        example: 1
      label:
        type: string
        example: "Consumer price inflation tables"
      dataset_id:
        type: string
        example: "consumer-price-inflation"
  MigrationJobList:
    allOf:
      - $ref: "#/definitions/List"
      - type: object
        properties:
          items:
            type: array
            description: Array containing results.
            items:
              $ref: "#/definitions/MigrationJob"
...
  List:
    type: object
    properties:
      count:
        type: integer
        description: How many jobs are present in the response
      limit:
        type: integer
        description: Max number of jobs we are returning with this response
      offset:
        type: integer
        description: The number of documents into the full list this response started at
      total_count:
        type: integer
        description: How many jobs are available in total
  MigrationState:
    type: string
    enum: *STATE
  MigrationJobType:
    type: string
    enum:
      - static_dataset
  MigrationTaskType:
    type: string
    enum:
      - dataset_series
      - dataset_edition
      - dataset_version
      - dataset_download
  ErrorList:
    type: object
    properties:
      errors:
        type: array
        items:
          $ref: '#/definitions/Error'
  Error:
    type: object
    properties:
      code:
        type: integer
        description: "Code representing the type of error that occurred"
      description:
        type: string
        description: "Description of the error"
securityDefinitions:
  Authorization:
    name: Authorization
    description: "Service Auth Token used by Auth service"
    in: header
    type: apiKey